+++ Code/xml/XmlPoint.java
@@ -6,13 +6,13 @@ import javax.xml.bind.annotation.XmlRootElement;
 @XmlRootElement(name = "XmlPoint")
 public class XmlPoint {
 
-	private double x;
-	private double y;
+	private int x;
+	private int y;
 	
 	/**
 	 * @return the x
 	 */
-	public double getX() {
+	public int getX() {
 		return x;
 	}
 	
@@ -20,14 +20,14 @@ public class XmlPoint {
 	 * @param x the x to set
 	 */
 	@XmlElement (name = "X")
-	public void setX(double x) {
+	public void setX(int x) {
 		this.x = x;
 	}
 	
 	/**
 	 * @return the y
 	 */
-	public double getY() {
+	public int getY() {
 		return y;
 	}
 	
@@ -35,7 +35,7 @@ public class XmlPoint {
 	 * @param y the y to set
 	 */
 	@XmlElement (name = "Y")
-	public void setY(double y) {
+	public void setY(int y) {
 		this.y = y;
 	}
